// client/src/generated/api/models/recommendation-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OrbiMed Analyst Trade Portal API
 * API for managing trade recommendations and tickets
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { SecurityResponse } from './security-response';
// May contain unused imports in some cases
// @ts-ignore
import type { StrategyResponse } from './strategy-response';

/**
 * 
 * @export
 * @interface RecommendationResponse
 */
export interface RecommendationResponse {
    /**
     * 
     * @type {number}
     * @memberof RecommendationResponse
     */
    'id': number;
    /**
     * 
     * @type {number}
     * @memberof RecommendationResponse
     */
    'analyst_id': number;
    /**
     * 
     * @type {number}
     * @memberof RecommendationResponse
     */
    'security_id': number;
    /**
     * 
     * @type {string}
     * @memberof RecommendationResponse
     */
    'trade_direction': string;
    /**
     * 
     * @type {string}
     * @memberof RecommendationResponse
     */
    'current_price': string | null;
    /**
     * 
     * @type {string}
     * @memberof RecommendationResponse
     */
    'target_price': string;
    /**
     * 
     * @type {string}
     * @memberof RecommendationResponse
     */
    'time_horizon': string;
    /**
     * 
     * @type {string}
     * @memberof RecommendationResponse
     */
    'expected_exit_date': string | null;
    /**
     * 
     * @type {number}
     * @memberof RecommendationResponse
     */
    'analyst_score': number;
    /**
     * 
     * @type {string}
     * @memberof RecommendationResponse
     */
    'notes': string | null;
    /**
     * 
     * @type {string}
     * @memberof RecommendationResponse
     */
    'status': string;
    /**
     * 
     * @type {boolean}
     * @memberof RecommendationResponse
     */
    'is_draft': boolean;
    /**
     * 
     * @type {string}
     * @memberof RecommendationResponse
     */
    'created_at': string;
    /**
     * 
     * @type {string}
     * @memberof RecommendationResponse
     */
    'updated_at': string;
    /**
     * 
     * @type {SecurityResponse}
     * @memberof RecommendationResponse
     */
    'security'?: SecurityResponse | null;
    /**
     * 
     * @type {Array<StrategyResponse>}
     * @memberof RecommendationResponse
     */
    'strategies'?: Array<StrategyResponse>;
}